Modified no float branch

+++ src/headers/mnist.h
@@ -8,18 +8,18 @@
 #include <stdint.h>
 
 typedef struct MnistImage{  //Structrue for Mnist Image.
-	int number_of_columns;             //Define the width of image. --- columns
-	int number_of_rows;                //Define the highth ot image. --- rows
+	int16_t number_of_columns;             //Define the width of image. --- columns
+	int16_t number_of_rows;                //Define the highth ot image. --- rows
 	uint8_t** image_data;      //Dynamic Define 2D Image Data Arrary. 
 }MnistImage;
 
 typedef struct MnistImageArray{ //Structrue for Mnist Images Array
-	int32_t number_of_images;             //Define Number of Images.    
+	int32_t number_of_images;     //Define Number of Images.    
 	MnistImage* image_point;      //Point for Images Array.
 }*ImageArray;                   //Images Array.
 
 typedef struct MnistLabel{  //Strutrue for Mnist Label
-	int label_length;                    //The length of Label 
+	int16_t label_length;        //The length of Label 
 	int8_t* label_data;         //The data of label
 }MnistLabel;
